move Tabs to text-labels.ts
