+++ e2e/app/page/workspace-data-page.ts
@@ -2,7 +2,7 @@ import ConceptDomainCard, { Domain } from 'app/component/concept-domain-card';
 import Link from 'app/element/link';
 import DataResourceCard from 'app/component/data-resource-card';
 import ClrIconLink from 'app/element/clr-icon-link';
-import { Language, MenuOption, ResourceCard } from 'app/text-labels';
+import { Language, MenuOption, ResourceCard, Tabs } from 'app/text-labels';
 import { Page } from 'puppeteer';
 import { makeRandomName } from 'utils/str-utils';
 import { waitForDocumentTitle, waitWhileLoading } from 'utils/waits-utils';
@@ -11,7 +11,7 @@ import CohortBuildPage from './cohort-build-page';
 import DatasetBuildPage from './dataset-build-page';
 import NotebookPage from './notebook-page';
 import WorkspaceAnalysisPage from './workspace-analysis-page';
-import WorkspaceBase, { Tabs } from './workspace-base';
+import WorkspaceBase from './workspace-base';
 import ConceptSetSearchPage from './conceptset-search-page';
 import { SaveOption } from 'app/modal/conceptset-save-modal';
 import ConceptSetActionsPage from './conceptset-actions-page';
